+++ ionics_fits/models/containers/__init__.py
@@ -2,3 +2,4 @@ from .aggregate_model import AggregateModel
 from .reparametrized_model import ReparametrizedModel
 from .repeated_model import RepeatedModel
 from .mapped_model import MappedModel
+from .scaled_model import ScaledModel