+++ docs/changes.md
@@ -10,7 +10,12 @@ Changes are divided into four categories:
 * Other - small changes, such as tweaks to internals, which most users will not need to be aware of
 * New models
 
-## 1.3.0.7
+## 1.3.0.9
+
+### Bug fixes
+* Fix calling ``func`` on ``RabiFlop``
+
+## 1.3.0.8
 
 Fitting to datasets with an arbitrary number x-axis dimensions is now supported!